bugfix process

+++ model/product.go
@@ -2,8 +2,8 @@ package model
 
 // Product data of struct
 type Product struct {
-	ID    int     `gorm:"id" json:"id"`
-	Name  string  `gorm:"name" json:"name"`
-	Qty   int     `gorm:"qty" json:"qty"`
-	Price float64 `gorm:"price" json:"price"`
+	ID    int    `gorm:"id" json:"id"`
+	Name  string `gorm:"name" json:"name"`
+	Qty   int    `gorm:"qty" json:"qty"`
+	Price int    `gorm:"price" json:"price"`
 }
